# Documents/SmartCliff/EMS_BACKEND/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/facilities/venues": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get All Venues",
                "operationId": "get_all_venues_facilities_venues_get",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Show venues only by the Names.",
                            "title": "Name"
                        },
                        "description": "Show venues only by the Names."
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Show venues only from this location (can match partly).",
                            "title": "Location"
                        },
                        "description": "Show venues only from this location (can match partly)."
                    },
                    {
                        "name": "min_capacity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Show venues with at least this much capacity.",
                            "title": "Min Capacity"
                        },
                        "description": "Show venues with at least this much capacity."
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Show only items with price less than or equal to this value.",
                            "title": "Max Price"
                        },
                        "description": "Show only items with price less than or equal to this value."
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Sort by this column name (like price, name, or capacity).",
                            "title": "Sort By"
                        },
                        "description": "Sort by this column name (like price, name, or capacity)."
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SortOrder",
                            "description": "Sort order — 'asc' for ascending, 'desc' for descending.",
                            "default": "asc"
                        },
                        "description": "Sort order — 'asc' for ascending, 'desc' for descending."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/facilities/bands": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get All Bands",
                "operationId": "get_all_bands_facilities_bands_get",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter bands by name.",
                            "title": "Name"
                        },
                        "description": "Filter bands by name."
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter bands by genre.",
                            "title": "Genre"
                        },
                        "description": "Filter bands by genre."
                    },
                    {
                        "name": "member_count",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter bands by member count.",
                            "title": "Member Count"
                        },
                        "description": "Filter bands by member count."
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Show only bands with price less than or equal to this value.",
                            "title": "Max Price"
                        },
                        "description": "Show only bands with price less than or equal to this value."
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Sort by this column name (like price, name, or member_count).",
                            "title": "Sort By"
                        },
                        "description": "Sort by this column name (like price, name, or member_count)."
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SortOrder",
                            "description": "Sort order — 'asc' or 'desc'.",
                            "default": "asc"
                        },
                        "description": "Sort order — 'asc' or 'desc'."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/facilities/decorations": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get All Decorations",
                "operationId": "get_all_decorations_facilities_decorations_get",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter decorations by name.",
                            "title": "Name"
                        },
                        "description": "Filter decorations by name."
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter decorations by type.",
                            "title": "Type"
                        },
                        "description": "Filter decorations by type."
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Show only decorations with price less than or equal to this value.",
                            "title": "Max Price"
                        },
                        "description": "Show only decorations with price less than or equal to this value."
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Sort by this column name (like price or name).",
                            "title": "Sort By"
                        },
                        "description": "Sort by this column name (like price or name)."
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SortOrder",
                            "description": "Sort order — 'asc' or 'desc'.",
                            "default": "asc"
                        },
                        "description": "Sort order — 'asc' or 'desc'."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/facilities/snacks": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get All Snacks",
                "operationId": "get_all_snacks_facilities_snacks_get",
                "parameters": [
                    {
                        "name": "price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter snacks by price.",
                            "title": "Price"
                        },
                        "description": "Filter snacks by price."
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Sort by this column name (like price or id).",
                            "title": "Sort By"
                        },
                        "description": "Sort by this column name (like price or id)."
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SortOrder",
                            "description": "Sort order — 'asc' or 'desc'.",
                            "default": "asc"
                        },
                        "description": "Sort order — 'asc' or 'desc'."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SortOrder": {
                "type": "string",
                "enum": [
                    "asc",
                    "desc"
                ],
                "title": "SortOrder"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}